+++ src/test/fuzz/script_parsing.cpp
@@ -2,11 +2,12 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <script/parsing.h>
 #include <test/fuzz/FuzzedDataProvider.h>
 #include <test/fuzz/fuzz.h>
-#include <util/spanparsing.h>
+#include <util/string.h>
 
-FUZZ_TARGET(spanparsing)
+FUZZ_TARGET(script_parsing)
 {
     FuzzedDataProvider fuzzed_data_provider(buffer.data(), buffer.size());
     const size_t query_size = fuzzed_data_provider.ConsumeIntegral<size_t>();
@@ -15,16 +16,16 @@ FUZZ_TARGET(spanparsing)
     const Span<const char> const_span{span_str};
 
     Span<const char> mut_span = const_span;
-    (void)spanparsing::Const(query, mut_span);
+    (void)script::Const(query, mut_span);
 
     mut_span = const_span;
-    (void)spanparsing::Func(query, mut_span);
+    (void)script::Func(query, mut_span);
 
     mut_span = const_span;
-    (void)spanparsing::Expr(mut_span);
+    (void)script::Expr(mut_span);
 
     if (!query.empty()) {
         mut_span = const_span;
-        (void)spanparsing::Split(mut_span, query.front());
+        (void)Split(mut_span, query.front());
     }
 }
