+++ src/test/util_tests.cpp
@@ -6,6 +6,7 @@
 #include <common/signmessage.h> // For MessageSign(), MessageVerify(), MESSAGE_MAGIC
 #include <hash.h> // For Hash()
 #include <key.h>  // For CKey
+#include <script/parsing.h>
 #include <sync.h>
 #include <test/util/random.h>
 #include <test/util/setup_common.h>
@@ -16,7 +17,6 @@
 #include <util/moneystr.h>
 #include <util/overflow.h>
 #include <util/readwritefile.h>
-#include <util/spanparsing.h>
 #include <util/strencodings.h>
 #include <util/string.h>
 #include <util/time.h>
@@ -1292,9 +1292,9 @@ static std::string SpanToStr(const Span<const char>& span)
     return std::string(span.begin(), span.end());
 }
 
-BOOST_AUTO_TEST_CASE(test_spanparsing)
+BOOST_AUTO_TEST_CASE(test_script_parsing)
 {
-    using namespace spanparsing;
+    using namespace script;
     std::string input;
     Span<const char> sp;
     bool success;
